Centralize optimistic collection actions
