--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/tasks/hooks/useTaskOptimisticActions/index.ts
@@ -1,1 +1,0 @@
-export { useTaskOptimisticActions } from "./useTaskOptimisticActions";
