--- apps/desktop/src/renderer/routes/_authenticated/hooks/useChatSessionOptimisticActions/index.ts
@@ -1,1 +1,0 @@
-export { useChatSessionOptimisticActions } from "./useChatSessionOptimisticActions";
