+++ apps/desktop/src/renderer/routes/_authenticated/hooks/useOptimisticCollectionActions/index.ts
@@ -1,0 +1,4 @@
+export {
+	type PersistableTransaction,
+	useOptimisticCollectionActions,
+} from "./useOptimisticCollectionActions";
