--- apps/desktop/src/renderer/routes/_authenticated/hooks/useOptimisticCollectionMutation/index.ts
@@ -1,4 +1,0 @@
-export {
-	type PersistableTransaction,
-	useOptimisticCollectionMutation,
-} from "./useOptimisticCollectionMutation";
