--- apps/desktop/src/renderer/routes/_authenticated/hooks/useV2ProjectOptimisticActions/index.ts
@@ -1,1 +1,0 @@
-export { useV2ProjectOptimisticActions } from "./useV2ProjectOptimisticActions";
