--- apps/desktop/src/renderer/routes/_authenticated/hooks/useV2WorkspaceOptimisticActions/index.ts
@@ -1,1 +1,0 @@
-export { useV2WorkspaceOptimisticActions } from "./useV2WorkspaceOptimisticActions";
